Create a new serializer for unshipped revenue, and put methods in correct places
