+++ app/controllers/api/v1/revenue/weekly_controller.rb
@@ -1,12 +1,6 @@
-module Api
-  module V1
-    module Revenue
-      class WeeklyController < ApplicationController
-        def index
-          invoices = Invoice.weekly_revenue
-          render(json: RevenueSerializer.format_json(invoices), status: 200)
-        end
-      end
-    end
+class Api::V1::Revenue::WeeklyController < ApplicationController
+  def index
+    invoices = Invoice.weekly_revenue
+    render(json: WeeklyRevenueSerializer.format_json(invoices), status: 200)
   end
 end
