+++ app/controllers/concerns/exception_handler.rb
@@ -14,15 +14,4 @@ module ExceptionHandler
   def object_not_found_error(id)
     { "message": "Couldn't find Merchant with 'id': #{id}"}
   end
-
-  def unshipped_invoices_error
-    { "data": {
-      "message": "request could not be completed",
-      "code": 400,
-      "errors": [
-        "wrong data type does not exist",
-      ]
-    }
-    }
-  end
 end
